% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/render.R
\name{DrawClassification}
\alias{DrawClassification}
\title{Draw a single classification}
\usage{
DrawClassification(classifications, subjects, n, pg.width, pg.height,
  before = NULL)
}
\arguments{
\item{classifications}{- list of classifications from
\code{\link{ReadClassifications}}}

\item{subjects}{- list of subjects from
\code{\link{ReadSubjects}}}

\item{pg.width}{- viewport width in pixels}

\item{pg.height}{- viewport height in pixels}

\item{before}{- POSIXt date-time, if not NULL (default), draw
only annotations from this time or earlier.}

\item{w}{index of classification to be drawn}
}
\description{
Call from inside a viewport of the size given in the call
}

