% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/render.R
\name{DrawLabel}
\alias{DrawLabel}
\title{Add a label to the presnet viewport.}
\usage{
DrawLabel(label)
}
\arguments{
\item{label}{- string used as label.}
}
\description{
Often the date-time, but can be anything.
}
\details{
In the bottom-right.
}

