% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/render.R
\name{DrawTranscription}
\alias{DrawTranscription}
\title{Draw a single transcription}
\usage{
DrawTranscription(n, pg.width, pg.height, before = NULL)
}
\arguments{
\item{pg.width}{- viewport width in pixels}

\item{pg.height}{- viewport height in pixels}

\item{before}{- POSIXt date-time, draw the transcription soonest
after this time in detail.}

\item{w}{index of classification to be drawn}
}
\description{
Call from inside a viewport of the size given in the call
}

