% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{FixStartAndFinish}
\alias{FixStartAndFinish}
\title{Make some more plausible start and end dates}
\usage{
FixStartAndFinish(classifications)
}
\arguments{
\item{classifications}{list from \code{\link{readClassifications}}.}
}
\value{
modified list with added $meta$new_started_at and $meta$new_finished_at components.
}
\description{
Some classifications have a very long gap between
start and end dates - estimate better start and end dates
from the annotation timestamps.
}

