% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{GetClassificationsByDate}
\alias{GetClassificationsByDate}
\title{Get classifications by date}
\usage{
GetClassificationsByDate(start_date, end_date, strict = TRUE)
}
\arguments{
\item{start_date}{POSIXt date for selection period start}

\item{end_date}{POSIXt date for selection period end}

\item{strict}{Boolean, if TRUE (default) check an annotation occured in the range.}
}
\value{
vector of indices of classifications in the selected date range.
}
\description{
Extract a subset of the classifications active between two given dates.
}
\details{
Optionally check that an annotation occurred between the two dates (slow).
}

