% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{GetPageImage}
\alias{GetPageImage}
\title{Get a raster image of a selected page}
\usage{
GetPageImage(i)
}
\arguments{
\item{i}{index of the selected subject}
}
\value{
raster of the image.
}
\description{
Download the image from the website (unless it's already cached).
}

