% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/render.R
\name{InterpolateLayout}
\alias{InterpolateLayout}
\title{Interpolate classification positions between two layouts}
\usage{
InterpolateLayout(old.layout, new.layout, fraction)
}
\arguments{
\item{old.layout}{- starting layout}

\item{new.layout}{- ending layout}

\item{fraction}{- 0-1, fraction of new layout}

\item{cls}{- vector of classification indices}
}
\value{
interpolated layout
}
\description{
When changing the number of pages shown at once, need
to shift between two layouts.
}

