% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/render.R
\name{Layout}
\alias{Layout}
\title{Layout n pages on the screen}
\usage{
Layout(x, y, n, aspect = 1.5, border = 0.05)
}
\arguments{
\item{x}{- screen width in pixels}

\item{y}{- screen height in pixels}

\item{n}{- how many pages to show at once}

\item{aspect}{- aspect ratio of page images.}

\item{border}{- fraction of page to be allocated
to border around each viewport}
}
\value{
list with components: nx and ny - number of
  pages wide by height; contents - vector length nx*ny
  showing which classification is in each viewport;
  viewports - list length nx*ny giving x,y, width and height
  for each page viewport.
}
\description{
Depends on page aspect ratio and screen size
}

