% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{MissingPageImage}
\alias{MissingPageImage}
\title{Mock a raster image of a selected page}
\usage{
MissingPageImage()
}
\value{
mock raster.
}
\description{
Fore use when the actual image is unavailable.
}

