% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oldWeather5.R
\docType{package}
\name{oldWeather5}
\alias{oldWeather5}
\alias{oldWeather5-package}
\title{oldWeather5 - Process data for Panoptes oldWeather}
\description{
Reads the data files downloaded from
 http://www.zooniverse.org/lab/1253/data-exports (project data) &
 http://www.zooniverse.org/lab/28/data-exports (Talk data).
}
\section{functions}{

\itemize{
  \item \code{\link{ReadClassifications}} - Get latest classifications dump
}
}

\section{data structure}{

 Records read in are put in a data frames, rows are records,
  columns are variables (YR, MO, SST, etc.)
}

