% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subjects.R
\name{ReadSubjects}
\alias{ReadSubjects}
\title{Read in the oW5 subjects table}
\usage{
ReadSubjects(file)
}
\arguments{
\item{file}{file to read data from.}
}
\value{
list of data frames.
}
\description{
Keeps the data internally in a data frame
}

