% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{SetIsTranscription}
\alias{SetIsTranscription}
\title{Is this in the annotation workflow, or transcription}
\usage{
SetIsTranscription(classifications)
}
\arguments{
\item{classifications}{list from \code{\link{ReadClassifications}}.}
}
\value{
modified list with added $meta$is_transcription and t components.
}
\description{
Search the annotations for 'contents'
}

