% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classifications.R
\name{UnpackMeta}
\alias{UnpackMeta}
\title{Make a data frame from the metadata classification strings.}
\usage{
UnpackMeta(meta)
}
\arguments{
\item{meta}{- Vector of JSON strings.}
}
\value{
data frame, name::value pairs from the JSON
}
\description{
Uses the rjson library.
}
\details{
Slow - explicit loops, not a problem with current data volumes.
}

