% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subjects.R
\name{UnpackSLocation}
\alias{UnpackSLocation}
\title{Make a data frame from the annotations location strings.}
\usage{
UnpackSLocation(location)
}
\arguments{
\item{location}{- Vector of JSON strings.}
}
\value{
data frame, name::value pairs from the JSON
}
\description{
Uses the rjson library.
}
\details{
Slow - explicit loops, not a problem with current data volumes.
}

