% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/render.R
\name{UpdateLayout}
\alias{UpdateLayout}
\title{Update a layout with a new set of classifications}
\usage{
UpdateLayout(old.layout, x, y, cls, aspect = 1.5, border = 0.05)
}
\arguments{
\item{old.layout}{- previous layout to be updated,
can be NULL, if so, start from scratch.}

\item{x}{- screen width in pixels}

\item{y}{- screen height in pixels}

\item{cls}{- vector of classification indices}

\item{aspect}{- aspect ratio of page images.}

\item{border}{- fraction of page to be allocated
to border around each viewport}
}
\value{
list with components: nx and ny - number of
  pages wide by height; contents - vector length nx*ny
  showing which classification is in each viewport;
  viewports - list length nx*ny giving x,y, width and height
  for each page viewport.
}
\description{
Keep the same classifications in the same viewports
  as far as possible
}

