% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/render.R
\name{DrawLayout}
\alias{DrawLayout}
\title{Draw a Layout}
\usage{
DrawLayout(layout, before = NULL)
}
\arguments{
\item{layout}{layout to be drawn \code{\link{Layout}}}

\item{before}{- POSIXt date-time, if not NULL (default), draw
only annotations from this time or earlier.}

\item{pg.width}{- viewport width in pixels}

\item{pg.height}{- viewport height in pixels}
}
\description{
Call from inside a viewport of the size given in the call
}

