% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classifications.R
\name{ReadClassifications}
\alias{ReadClassifications}
\title{Read in the oW5 classifications table}
\usage{
ReadClassifications(file)
}
\arguments{
\item{file}{file to read data from.}
}
\value{
data frame - 1 row per record, column names as in the file.
}
\description{
Keeps the data internally in a data frame
}

